cleanup

+++ linux_driver/mymodule/xbasic_types.h
@@ -38,15 +38,15 @@ extern void XAssert(char *, int);
  * Deprecated legacy types.
  * @{
  */
-typedef unsigned char Xuint8;   /**< unsigned 8-bit */
-typedef char        Xint8;    /**< signed 8-bit */
-typedef unsigned short  Xuint16;  /**< unsigned 16-bit */
-typedef short       Xint16;   /**< signed 16-bit */
-typedef unsigned long Xuint32;  /**< unsigned 32-bit */
-typedef long        Xint32;   /**< signed 32-bit */
-typedef float       Xfloat32; /**< 32-bit floating point */
-typedef double        Xfloat64; /**< 64-bit double precision FP */
-typedef unsigned long Xboolean; /**< boolean (XTRUE or XFALSE) */
+typedef unsigned char  Xuint8;   /**< unsigned 8-bit */
+typedef char           Xint8;    /**< signed 8-bit */
+typedef unsigned short Xuint16;  /**< unsigned 16-bit */
+typedef short          Xint16;   /**< signed 16-bit */
+typedef unsigned       Xuint32;  /**< unsigned 32-bit */
+typedef int            Xint32;   /**< signed 32-bit */
+typedef float          Xfloat32; /**< 32-bit floating point */
+typedef double         Xfloat64; /**< 64-bit double precision FP */
+typedef unsigned       Xboolean; /**< boolean (XTRUE or XFALSE) */
 
 typedef struct
 {
